% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{calc_exp_totals}
\alias{calc_exp_totals}
\title{Total exposed}
\usage{
calc_exp_totals(l_out)
}
\arguments{
\item{l_out}{List with output from SC-COSMO and all parameters}
}
\value{
A data.frame with the total number of exposed as columns over time
}
\description{
\code{calc_exp_totals} Calculate total number of exposed over time.
}
