% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{calc_inf_nodx}
\alias{calc_inf_nodx}
\title{Undetected infections}
\usage{
calc_inf_nodx(l_out)
}
\value{
A data.frame with the number of undetected infections as columns over time
}
\description{
\code{calc_inf_nodx} Calculate number of undetected infections over time.
}
