% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{calc_inf_totals}
\alias{calc_inf_totals}
\title{Total infections}
\usage{
calc_inf_totals(l_out)
}
\value{
A data.frame with the total number of infections as columns over time
}
\description{
\code{calc_inf_totals} Calculate total number of infections over time.
}
