% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{compute_inflows}
\alias{compute_inflows}
\title{Compute inflows considering competing risks using convolution of binomials
\code{compute_inflows} computes inflows}
\usage{
compute_inflows(
  n_hhsize,
  n_hh_mod,
  v_names_hh,
  v_names_states,
  v_names_source,
  v_names_destiny,
  v_names_source_destiny,
  v_index_hh_source,
  v_index_hh_destiny,
  v_index_keep_source,
  v_index_souce_destiny_hh,
  m_possibilities,
  r_flow,
  n_source_states,
  m_hh_out
)
}
\value{
A matrix with inflow rates.
}
\description{
Compute inflows considering competing risks using convolution of binomials
\code{compute_inflows} computes inflows
}
