% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{daily_rate}
\alias{daily_rate}
\title{Daily rate}
\usage{
daily_rate(cum_prop, duration)
}
\arguments{
\item{cum_prop}{Cumulative proportion}

\item{duration}{total number of days to achieve proportion}
}
\value{
A daily rate
}
\description{
\code{daily_rate} generates the daily rate to achieve a cumulative proportion
over a given duration.
}
