% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{gen_household_matrices_mc_seir}
\alias{gen_household_matrices_mc_seir}
\title{Generate household matrices}
\usage{
gen_household_matrices_mc_seir(
  n_hhsize,
  n_hh_mod,
  v_names_states,
  v_names_exp,
  v_names_inf,
  v_names_rec,
  v_names_vax,
  v_index_hh_sus,
  v_index_hh_exp,
  v_index_hh_inf,
  v_index_hh_rec,
  v_index_hh_vax,
  df_possibilities,
  r_sigma,
  r_gamma,
  r_omega,
  r_vax_omega
)
}
\arguments{
\item{n_hhsize}{Household size}

\item{n_hh_mod}{Number of household model states}

\item{v_names_states}{Vector with all names}

\item{v_names_exp}{Vector with names of exposed states}

\item{v_names_inf}{Vector with names of infectious states}

\item{v_names_rec}{Vector with names of recovered states}

\item{v_names_vax}{Vector with names of vaccination states}

\item{v_index_hh_sus}{Indexing column vector of susceptibles}

\item{v_index_hh_exp}{Indexing column vector of exposed}

\item{v_index_hh_inf}{Indexing column vector of infectious}

\item{v_index_hh_rec}{Indexing column vector of recovered}

\item{v_index_hh_vax}{Indexing column vector of vaccination}

\item{r_sigma}{Daily rate of progression of exposed individuals
(Latent period)}

\item{r_gamma}{Daily rate of recovery of infectious individuals
(Infectiousness period).}

\item{r_omega}{Waning rate for the household model members within each
epidemic compartment}

\item{r_vax_omega}{Vaccination waning rate for the household model members}

\item{m_possibilities}{Matrix with possible combinations of household
members within each epidemic compartment}
}
\value{
A list with three transition matrices: 1) community transmission matrix; 2)
household transmission matrix; and 3) household recovered matrix
}
\description{
\code{gen_household_matrices} generates transition matrices for within
household epidemics.
}
