% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_npi}
\alias{get_npi}
\title{Get non-pharmaceutical intervention (NPI) for a given time}
\usage{
get_npi(n_time, parameters)
}
\arguments{
\item{n_time}{A time for which to retrieve the NPI level}

\item{parameters}{List with all parameters of model}
}
\value{
A value of the NPI level.
}
\description{
\code{get_npi} is used to get the time-varying NPI effect
}
