% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_vax}
\alias{get_vax}
\title{Get vaccination (Vax) rate for a given time}
\usage{
get_vax(n_time, parameters)
}
\arguments{
\item{n_time}{A time for which to retrieve the Vax rate}

\item{parameters}{List with all parameters of model}
}
\value{
A value of the Vax rate.
}
\description{
\code{get_vax} is used to get the time-varying vaccination rate
}
