% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh-mc-seirv.R
\docType{package}
\name{hh-mc-seirv}
\alias{hh-mc-seirv}
\title{\code{hh-mc-seirv} package}
\description{
Household multi-compartment (MC) susceptible-exposed-infectious-recovered-vaccinated (SEIRV) model
}
\details{
See the description of the package on
\href{https://SC-COSMO.github.io/hh-mc-seirv/index.html}{darthpack website}
or the README on
\href{https://github.com/SC-COSMO/hh-mc-seirv#readme}{GitHub}
}
