% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{hh_mc_seir}
\alias{hh_mc_seir}
\title{The household model for a multi-compartment SEIR (HH-MC_SEIR) epidemic model.}
\usage{
hh_mc_seir(l_params_all)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model}
}
\value{
A data.frame with population for each class over time.
}
\description{
\code{hh_mc_seir_out} implements the household model for a multi-compartment
SEIR (HH-MC_SEIR) epidemic model.
}
