% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{hh_mc_seir_dx}
\alias{hh_mc_seir_dx}
\title{Derivatives of the household model for a multi-compartment SEIR (HH-MC_SEIR)
epidemic model.}
\usage{
hh_mc_seir_dx(t, state, parameters)
}
\arguments{
\item{time}{Time at which the derivatives will be computed}

\item{v_pop}{Vector with the population by class}

\item{l_params_all}{List with all parameters of SC-COSMO model}
}
\value{
A list with a vector of derivatives for each class .
}
\description{
\code{hh_mc_seir_dx} computes the derivatives of the household model for
a multi-compartment SEIR (HH-MC_SEIR) epidemic model.
}
