% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhmcseirv.R
\docType{package}
\name{hhmcseirv}
\alias{hhmcseirv}
\title{\code{hhmcseirv} package}
\description{
Household multi-compartment (MC) susceptible-exposed-infectious-recovered-vaccinated (SEIRV) model
}
\details{
See the description of the package on
\href{https://SC-COSMO.github.io/hhmcseirv/index.html}{darthpack website}
or the README on
\href{https://github.com/SC-COSMO/hhmcseirv#readme}{GitHub}
}
