% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{number_of_hh_odes}
\alias{number_of_hh_odes}
\title{Number of household ordinary differential equations (ODEs)}
\usage{
number_of_hh_odes(n_states, n_hhsize)
}
\arguments{
\item{n_states}{Number of health states}

\item{n_hhsize}{Household size}
}
\value{
The number of household ordinary differential equations (ODEs) for a given
household size and number of health states
}
\description{
\code{number_of_odes} generates the daily rate to achieve a cumulative proportion
over a given duration.
}
